[#922] Extract input field to a component
with @anikarni
