+++ web-ui/src/account_recovery/page.scss
@@ -60,50 +60,6 @@ a {
 
 }
 
-.field-group {
-  position:relative;
-  margin: 1.5em 0;
-  width: 100%;
-
-  label {
-    font-size: 0.9em;
-    margin-bottom: 10px;
-    display: inline-block;
-  }
-
-  input {
-    display: block;
-    border: none;
-    border-bottom: solid 1px #4da3b6;
-    width: 100%;
-    height: auto;
-    padding: 0.3em 0px;
-  }
-
-  input:focus {
-    outline:none;
-    border-width: 2px;
-  }
-
-  input:focus ~ .animated-label, input:valid ~ .animated-label {
-    top:-12px;
-    left: 0;
-    font-size:0.7em;
-    color:#4da3b6;
-  }
-
-  .animated-label {
-    color:#999;
-    position:absolute;
-    pointer-events:none;
-    left: 6px;
-    top:10px;
-    transition:0.2s ease all;
-    -moz-transition:0.2s ease all;
-    -webkit-transition:0.2s ease all;
-  }
-}
-
 .container {
   background: $white;
   width: 84%;
@@ -135,7 +91,7 @@ img {
     -webkit-flex-direction: column;
     flex-direction: column;
 
-    .field-group, div {
+    div {
       flex-basis: 70%;
       -webkit-align-self: center;
       align-self: center;
@@ -157,15 +113,10 @@ img {
   form {
     margin-left: 2.5em;
 
-    .field-group, div {
+    div {
       width: 300px;
       -webkit-align-self: flex-start;
       align-self: flex-start;
-    }
-
-    .field-group {
-      margin-top: 3em;
-      font-size: 1em;
     }
   }
