+++ web-ui/src/common/submit_button/submit-button.scss
